% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bettivector.R
\name{betti.fun}
\alias{betti.fun}
\title{compute the difference between birth-death at time t from a matrix.
internal function used in bettimatrix}
\usage{
betti.fun(bcmat, t)
}
\arguments{
\item{bcmat}{is a ?times? matrix}

\item{t}{is the time or distance to compute betti.fun}
}
\value{
the number of birth - death
}
\description{
compute the difference between birth-death at time t from a matrix.
internal function used in bettimatrix
}
