% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bettivector.R
\name{bettimatrix}
\alias{bettimatrix}
\title{Compute Betti Curves}
\usage{
bettimatrix(diag, order = -1, xmax = -1, l = 100)
}
\arguments{
\item{diag}{is a persistence diagram from TDA o TDAstats}

\item{order}{is the maximum order the betti curve is computed}

\item{xmax}{is the maximum length/time the betti curve is computed}

\item{l}{is the dimension of each betticurve}
}
\value{
a 3 times l matrix with the betticurve
}
\description{
Compute Betti Curves until some order from persistence diagram
if order=-1 then the order is infered from diag
if xmax=-1 the max lenght is infered from diag
}
\examples{
library(TDA)
data <- cbind(runif(10), runif(10))
alpha_complex <- alphaComplexDiag(data)
alpha_diag <- alpha_complex$diagram
bettimatrix(alpha_complex)
}
