% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{compute_betti_curves}
\alias{compute_betti_curves}
\title{Compute Betti curves}
\usage{
compute_betti_curves(diag, maxdim = -1, maxradius = -1, len = 100)
}
\arguments{
\item{diag}{is a persistence diagram from TDA o TDAstats}

\item{maxdim}{is the maximum dimension the betti curve is computed}

\item{maxradius}{is the maximum radius the betti curve is computed}

\item{len}{is the length of each betticurve}
}
\value{
A matrix with the Betti curves
}
\description{
Compute Betti curves of a persistence diagram
until a desired dimension, a maximum radius, and output vector lenght.
The Betti curve of dimension \code{d} of a diagram \code{diag}
refers to the sequence of Betti numbers of dimension \code{d} and
diamgram \code{diag} along radius from zero to infinity (maximun radius).
}
\details{
if \code{maxdim = -1} then the dimension is infered from diag
if \code{maxradius = -1} the maximum radius is infered from diag
}
\examples{
library(TDAstats)
X <- cbind(runif(100), runif(100))
diag <- calculate_homology(X)
compute_betti_curves(diag)
}
