% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{compute_betti_number}
\alias{compute_betti_number}
\title{Compute the Betti number}
\usage{
compute_betti_number(diag, dim, radius)
}
\arguments{
\item{diag}{is a persistence diagram generated by a point cloud}

\item{dim}{is the dimension of the homology group}

\item{radius}{is the radius for which the betti number is computed}
}
\value{
the betti number (birth - death)
}
\description{
Compute the Betti number of a persistence diagram
for a desired dimension and radius.
The Betti number of dimension \code{d} and radius \code{r}
of a diagram \code{diag}
refers to the number of \code{d}-dimensional holes of
a simplicial complex at radius \code{r} of
a \code{diag} generated by a point cloud.
When \code{d = 0} it should be interpreted as the
number of conected components of the simplicial complex
at radius \code{r}.
}
