% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{plot_betti_curve}
\alias{plot_betti_curve}
\title{Plot Betti Curves}
\usage{
plot_betti_curve(betticurve)
}
\arguments{
\item{betticurve}{is a betti matrix from bettimatrix}
}
\description{
Plot Betti Curves from persistence diagram of desired dimension
in a ggplot style with facets and free scales
}
\examples{
plot_betti(betticurve)
}
